+++ PSModuleDevelopment/functions/assembly/Find-PSMDType.ps1
@@ -29,6 +29,9 @@
 	
 	.PARAMETER Enum
 		Whether the type to find must be an enumeration.
+		
+	.PARAMETER Static
+		Whether the type to find must be static.
 	
 	.PARAMETER Implements
 		Whether the type to find must implement this interface
@@ -70,6 +73,9 @@
 		[switch]
 		$Enum,
 		
+		[switch]
+		$Static,
+		
 		[string]
 		$Implements,
 		
@@ -84,6 +90,7 @@
 	{
 		$boundEnum = Test-PSFParameterBinding -ParameterName Enum
 		$boundPublic = Test-PSFParameterBinding -ParameterName Public
+		$boundStatic = Test-PSFParameterBinding -ParameterName Static
 	}
 	process
 	{
@@ -114,6 +121,7 @@
 				if ($boundEnum -and ($Enum -ne $type.IsEnum)) { continue }
 				if ($InheritsFrom -and ($type.BaseType.FullName -notlike $InheritsFrom)) { continue }
 				if ($Attribute -and ($type.CustomAttributes.AttributeType.Name -notlike $Attribute)) { continue }
+				if ($boundStatic -and ($Static -ne ($type.IsAbstract -and $type.IsSealed))) { continue }
 				
 				$type
 			}