+++ templates/PSFCimFunction/þnameþ.ps1
@@ -117,8 +117,4 @@
 		}
 		#endregion Process by Computer Name
 	}
-	end
-	{
-		
-	}
 }