+++ templates/PSFModule/internal/scripts/license.ps1
@@ -1,4 +1,4 @@
-﻿New-PSFLicense -Product 'þnameþ' -Manufacturer 'þauthorþ' -ProductVersion $PSModuleVersion -ProductType Module -Name MIT -Version "1.0.0.0" -Date (Get-Date "þ!date!þ") -Text @"
+﻿New-PSFLicense -Product 'þnameþ' -Manufacturer 'þauthorþ' -ProductVersion $script:ModuleVersion -ProductType Module -Name MIT -Version "1.0.0.0" -Date (Get-Date "þ!date!þ") -Text @"
 Copyright (c) þ!year!þ þauthorþ
 
 Permission is hereby granted, free of charge, to any person obtaining a copy