+++ templates/PSFProject/PSMDTemplate.ps1
@@ -1,11 +1,11 @@
-@{
+﻿@{
 	TemplateName = 'PSFProject'
-	Version = "1.0.0.0"
+	Version = "1.1.1.0"
 	AutoIncrementVersion = $true
 	Tags = 'module','psframework'
 	Author = 'Friedrich Weinmann'
 	Description = 'PowerShell Framework based project scaffold'
-	Exclusions = @("PSMDInvoke.ps1") # Contains list of files - relative path to root - to ignore when building the template
+	Exclusions = @("PSMDInvoke.ps1", ".PSMDDependency") # Contains list of files - relative path to root - to ignore when building the template
 	Scripts = @{
 		guid = {
 			[System.Guid]::NewGuid().ToString()
@@ -24,6 +24,9 @@
 		}
 		guid4 = {
 			[System.Guid]::NewGuid().ToString().ToUpper()
+		}
+		psframework = {
+			(Get-Module PSFramework).Version.ToString()
 		}
 	}
 }